+++ sdk/storage/azure-storage-blobs/vcpkg.json
@@ -16,6 +16,9 @@
     {
       "name": "vcpkg-cmake-config",
       "host": true
+    },
+    {
+      "name": "nanoarrow"
     }
   ]
 }
