+++ sdk/storage/azure-storage-blobs/vcpkg/Config.cmake.in
@@ -5,6 +5,7 @@
 
 include(CMakeFindDependencyMacro)
 find_dependency(azure-storage-common-cpp)
+find_dependency(nanoarrow)
 
 include("${CMAKE_CURRENT_LIST_DIR}/azure-storage-blobs-cppTargets.cmake")
