+++ sdk/storage/azure-storage-blobs/vcpkg/vcpkg.json
@@ -27,6 +27,9 @@
     {
       "name": "vcpkg-cmake-config",
       "host": true
+    },
+    {
+      "name": "nanoarrow"
     }
   ]
 }
